# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "url": "https://esocialbrasil.com.br/",
            "email": "exemplo@exemplo.com.br"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/": {
            "get": {
                "description": "Home page",
                "operationId": "58658e34a5c8bd71c4a0afed6de10d20",
                "responses": {
                    "default": {
                        "description": "Benvindo ao site"
                    }
                }
            }
        },
        "/api/tarefa": {
            "get": {
                "tags": [
                    "Tarefa"
                ],
                "summary": "Retornar uma lista das tarefas",
                "description": "Retornar os objetos dos tarefas",
                "operationId": "a79e012fce138fc62b731aae394afd11",
                "responses": {
                    "200": {
                        "description": "Uma lista com tarefas"
                    },
                    "404": {
                        "description": "Nenhuma lista de tarefas encontrada"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        },
        "/api/tarefa/{id}": {
            "put": {
                "tags": [
                    "Tarefa"
                ],
                "summary": "Edita uma tarefa",
                "description": "Retornar os objetos da tarefa",
                "operationId": "67cd449521a31562fafdca4c490d2ef8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id da tarefa",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "descricao",
                        "in": "query",
                        "description": "Descrição da tarefa",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "str"
                        }
                    },
                    {
                        "name": "tipo_desenvolvimento",
                        "in": "query",
                        "description": "Tipo de desenvolvimento",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "Backend",
                            "enum": [
                                "Backend",
                                "Frontend",
                                "Database"
                            ]
                        }
                    },
                    {
                        "name": "nivel_dificuldade",
                        "in": "query",
                        "description": "Nível de dificuldade",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "Moderada",
                            "enum": [
                                "Fácil",
                                "Moderada",
                                "Difícil"
                            ]
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status da tarefa",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "Aberta",
                            "enum": [
                                "Aberta",
                                "Fechada",
                                "Cancelada"
                            ]
                        }
                    },
                    {
                        "name": "concluida",
                        "in": "query",
                        "description": "Tarefa concluída",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    },
                    {
                        "name": "responsavel",
                        "in": "query",
                        "description": "Responsável pela tarefa (ID selecionado da tabela de usuários)",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                1,
                                2,
                                3,
                                4
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarefa salva com sucesso"
                    },
                    "422": {
                        "description": "Atributos não podem ser processados"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    },
                    "400": {
                        "description": "Erro na solicitação"
                    },
                    "401": {
                        "description": "Não autorizado, autenticação requerida"
                    },
                    "403": {
                        "description": "Acesso proibido"
                    },
                    "404": {
                        "description": "Recurso não encontrado"
                    },
                    "409": {
                        "description": "Conflito ao processar a solicitação"
                    },
                    "429": {
                        "description": "Muitas solicitações, tente novamente mais tarde"
                    }
                }
            }
        },
        "/api/user": {
            "get": {
                "tags": [
                    "Usuário"
                ],
                "summary": "Retornar uma lista de usuários",
                "description": "Retornar os objetos dos usuários",
                "operationId": "971db9bc974c0e607288cfe7ecc10096",
                "responses": {
                    "200": {
                        "description": "Uma lista com usuários"
                    },
                    "404": {
                        "description": "Nenhuma lista de usuários encontrada"
                    },
                    "500": {
                        "description": "Erro interno do servidor"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Tarefa",
            "description": "Tarefa"
        },
        {
            "name": "Usuário",
            "description": "Usuário"
        }
    ]
}